+++ SENAI_HROADS_TARDE/Projeto/senai.hroads.webApi/senai.hroads.webApi/Interfaces/IClasseRepository.cs
@@ -8,7 +8,7 @@ namespace senai.hroads.webApi_.Interfaces
         List<Classe> Listar();
         Classe BuscarPorId(int idClasse);
         void Cadastrar(Classe novoClasse);
-        void Atualizar(int idClasse, Classe classeAtualizado);
+        void Atualizar(byte idClasse, Classe classeAtualizado);
         void Deletar(int idClasse);
     }
 }
